LipSyncStudio/requirements.txt
opencv-python
librosa
diffusers[torch]
TTS
openai-whisper
mediapipe